requirements.txt
fastapi
uvicorn
sqlalchemy
streamlit
pandas
transformers
torch
accelerate
bitsandbytes
python-dotenv
plotly
pydantic
aiosqlite
sqlparse
numpy
ctransformers[cuda]
tqdm
requests
SpeechRecognition
pyttsx3
PyAudio
# Database drivers
psycopg2-binary
pyodbc
# Alternative PostgreSQL driver (if psycopg2 fails)
pg8000
# MySQL driver
pymysql
# Oracle driver
cx-oracle
# Advanced data analysis
scipy
seaborn
matplotlib
openpyxl
altair
# Additional visualization libraries
bokeh
holoviews
# Statistical analysis
statsmodels
scikit-learn